Moved RoundManager into its own folder/package
